# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trosc API Documentation",
    "version": "1.0.0",
    "description": "Official API documentation for the Trosc platform.",
    "contact": {
      "name": "Basem Esam",
      "url": "https://github.com/basem3sam"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/v1",
      "description": "development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Event": {
        "type": "object",
        "required": [
          "title",
          "description",
          "date",
          "locationType",
          "createdBy"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "6713b5ac12ef4567890a4444"
          },
          "title": {
            "type": "string",
            "description": "Event title",
            "example": "Backend Bootcamp: Introduction to Node.js"
          },
          "description": {
            "type": "string",
            "description": "Detailed event content",
            "example": "A 3-day backend development workshop covering Node.js, Express, and MongoDB."
          },
          "track": {
            "type": "string",
            "description": "Optional ObjectId reference to a Track",
            "example": "67123abc12ef4567890a1234"
          },
          "course": {
            "type": "string",
            "description": "Optional ObjectId reference to a Course",
            "example": "67123abc12ef4567890a5678"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time of the event",
            "example": "2025-11-10T10:00:00.000Z"
          },
          "locationType": {
            "type": "string",
            "enum": [
              "online",
              "offline"
            ],
            "description": "Whether event is online or physical",
            "example": "online"
          },
          "locationLink": {
            "type": "string",
            "description": "Zoom or meeting link for online events",
            "example": "https://zoom.us/meeting/xyz"
          },
          "locationAddress": {
            "type": "string",
            "description": "Address or venue for offline events",
            "example": "ITI Smart Village, Cairo"
          },
          "coverImage": {
            "type": "string",
            "description": "Event cover image",
            "example": "bootcamp-cover.jpg"
          },
          "attendees": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "User ObjectId"
            },
            "example": [
              "67123abc12ef4567890a2222"
            ]
          },
          "createdBy": {
            "type": "string",
            "description": "ObjectId reference to admin/instructor who created the event",
            "example": "67123abc12ef4567890a9999"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-18T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-18T15:00:00.000Z"
          }
        }
      },
...
      "StandardError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Error description here"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "getAllEvents",
        "summary": "Get all events",
        "description": "Retrieve all events with filtering, sorting, and pagination",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "locationType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "offline"
              ]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "-date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "events": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Event"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createEvent",
        "summary": "Create a new event",
        "description": "Create an event (admin and instructors only)",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              },
              "example": {
                "title": "Backend Bootcamp",
                "description": "3-day Node.js workshop",
                "date": "2025-11-10T10:00:00.000Z",
                "locationType": "online",
                "locationLink": "https://zoom.us/meeting/xyz"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/events/my-events": {
      "get": {
        "operationId": "getMyEvents",
        "summary": "Get my RSVP'd events",
        "description": "Retrieve all events the current user has RSVP'd to",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "events": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Event"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "getEvent",
        "summary": "Get single event",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "event": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateEvent",
        "summary": "Update event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteEvent",
        "summary": "Delete event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/events/{id}/rsvp": {
      "post": {
        "operationId": "rsvpEvent",
        "summary": "RSVP to an event",
        "description": "Register attendance for an event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RSVP confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "RSVP confirmed"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "event": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Already RSVP'd"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "cancelRsvp",
        "summary": "Cancel RSVP",
        "description": "Remove attendance registration from an event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RSVP cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "RSVP cancelled"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "event": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Not attending this event"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
...
}